# internal/admin/api/openapi.yaml
openapi: "3.0.3"
info:
  title: GoProxify Admin API
  description: API REST du Control Plane GoProxify
  version: "1.0"
servers:
  - url: /
    description: Instance courante
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gpx_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        created_at:
          type: string
          format: date-time
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/nodes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [nodes]
      summary: Détail d'un nœud
      responses:
        "200":
          description: OK
    delete:
      tags: [nodes]
      summary: Supprimer un nœud
      responses:
        "204":
          description: Supprimé
...
  /api/v1/users:
    get:
      tags: [users]
      summary: Liste des utilisateurs (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags: [users]
      summary: Créer un utilisateur (admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [admin, operator, viewer]
      responses:
        "201":
          description: Créé
  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      tags: [users]
      summary: Mettre à jour un utilisateur
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
    delete:
      tags: [users]
      summary: Supprimer un utilisateur
      responses:
        "204":
          description: Supprimé
...